+++ common/src/main/kotlin/net/eve0415/mc/WebsocketIntegration/Enum/WSIPlatformType.kt
@@ -1,0 +1,7 @@
+package net.eve0415.spigot.WebsocketIntegration.Enum
+
+enum class WSIPlatformType(val value: String) {
+  Paper("PaperMC"),
+  Sponge("Sponge"),
+  Velocity("Velocity")
+}
