+++ common/src/main/kotlin/net/eve0415/mc/WebsocketIntegration/Interface/ServerStatusInfo.kt
@@ -1,4 +1,4 @@
-package net.eve0415.spigot.WebsocketIntegration.Util
+package net.eve0415.spigot.WebsocketIntegration.Interface
 
 public interface ServerStatusInfo {
   fun getOnlinePlayers(): Int
