+++ common/src/main/kotlin/net/eve0415/mc/WebsocketIntegration/Interface/WSIChatHandler.kt
@@ -1,4 +1,4 @@
-package net.eve0415.spigot.WebsocketIntegration.Util
+package net.eve0415.spigot.WebsocketIntegration.Interface
 
 public interface WSIChatHandler {
   fun chatHandler(name: String, uuid: String, url: String, message: String)
