--- common/src/main/kotlin/net/eve0415/mc/WebsocketIntegration/Util/WSIConfiguration.kt
@@ -1,16 +1,0 @@
-package net.eve0415.spigot.WebsocketIntegration.Util
-
-public interface WSIConfiguration {
-  val address: String
-  val port: Int
-  val id: Int
-  val name: String
-
-  // fun getAddress(): String
-
-  // fun getPort(): Int
-
-  // fun getId(): String
-
-  // fun getName(): String
-}
