--- common/src/main/kotlin/net/eve0415/mc/WebsocketIntegration/Util/WSIPlatformType.kt
@@ -1,7 +1,0 @@
-package net.eve0415.spigot.WebsocketIntegration.Util
-
-enum class WSIPlatformType(val readable: String) {
-  Paper("PaperMC"),
-  Sponge("Sponge"),
-  Velocity("Velocity")
-}
