--- common/src/main/resources/assets/websocketintegration/config.yml
@@ -1,11 +1,0 @@
-# Address and Port to connect to Websocket
-address: localhost
-port: 25500
-
-# Server ID to manage your multiple servers.
-# Set id to 0 to use server port as server id.
-id: 0
-
-# auto will use the platform type as server name
-# Recommend to set this when you have specified id
-name: auto
