+++ settings.gradle
@@ -1,6 +1,6 @@
 pluginManagement {
   plugins {
-    id 'java'
+    id "java"
     id "org.jetbrains.kotlin.jvm" version "1.5.10"
     id "com.github.johnrengelman.shadow" version "7.0.0"
     id "net.kyori.blossom" version "1.2.0"
